+++ app/presenters/vendor_api/single_application_presenter.rb
@@ -26,6 +26,7 @@ module VendorAPI
             last_name: application_form.last_name,
             date_of_birth: application_form.date_of_birth,
             nationality: nationalities,
+            domicile: application_form.country,
             uk_residency_status: uk_residency_status,
             english_main_language: application_form.english_main_language,
             english_language_qualifications: application_form.english_language_details,
@@ -207,7 +208,7 @@ module VendorAPI
         institution_details: institution_details(qualification),
         awarding_body: qualification.awarding_body,
         equivalency_details: qualification.equivalency_details,
-      }
+      }.merge HesaQualificationFieldsPresenter.new(qualification).to_hash
     end
 
     def institution_details(qualification)