+++ app/views/api_docs/pages/release_notes.md
@@ -1,3 +1,10 @@
+### 6th October 2020
+
+New attributes:
+
+- `Candidate` now has a `domicile` attribute of type string, returning a two-letter country code (ISO 3166-2). Its value is derived from the candidate's address.
+- `Qualification` now includes attributes for HESA qualification codes, if the qualification is degree-level. The attributes are: `hesa_degtype`, `hesa_degsbj`, `hesa_degclss`, `hesa_degest`, `hesa_degctry`, `hesa_degstdt`, `hesa_degenddt`.
+
 ### 5th October 2020
 
 Changes to existing attributes: